refactor: split backup package by responsibility
